move parseTriggerTime to Event static function
